creating dynamic cultural tour
